requirements.txt
cffi==1.2.1
chardet==2.3.0
cryptography==1.9
dateutils==0.6.6
diff-match-patch==20121119
Django==1.8.3
django-constance==1.2.1
django-countries==3.3
django-extra-fields==0.3
django-import-export==0.5.1
django-model-utils==2.3.1
django-multiselectfield==0.1.5
django-notifications-hq==0.7.0
django-paginator-rawqueryset==0.1.0
django-password-reset==0.7
django-redis==4.3.0
django-tables2==1.4.2
djangorestframework==3.1.3
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.2
facebook-sdk==0.4.0
facepy==1.0.8
foursquare==1!2015.5.26
gnureadline==6.3.3
gunicorn==19.3.0
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
ipython==3.2.1
jsonfield==1.0.3
MySQL-python==1.2.5
paramiko==1.15.2
paypalrestsdk==1.11.0
Pillow==2.9.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyfcm==1.2.1
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.8.1
Sendinblue==2.0.5.1
six==1.9.0
stripe==1.48.1
tablib==0.11.2
ujson==1.33
django-crontab==0.7.1
xlwt
django-admin-list-filter-dropdown
django-categories
